+++ src/components/checkbox/index.ts
@@ -1,0 +1,2 @@
+export { default as UiCheckbox } from './UiCheckbox.vue';
+export type { UiCheckboxProps } from './UiCheckbox.types';